Split primitives.ts into multiple files
