+++ src/geo/projection/globe_covering_tiles.test.ts
@@ -1,4 +1,4 @@
-import {Aabb} from '../../util/primitives';
+import {Aabb} from '../../util/primitives/aabb';
 import {expectToBeCloseToArray} from '../../util/test/util';
 import {GlobeCoveringTilesDetailsProvider} from './globe_covering_tiles_details_provider';
