+++ src/geo/projection/mercator_transform.ts
@@ -15,7 +15,7 @@ import {EXTENT} from '../../data/extent';
 import type {ProjectionData, ProjectionDataParams} from './projection_data';
 import {scaleZoom, TransformHelper, zoomScale} from '../transform_helper';
 import {MercatorCoveringTilesDetailsProvider} from './mercator_covering_tiles_details_provider';
-import {Frustum} from '../../util/primitives';
+import {Frustum} from '../../util/primitives/frustum';
 import {CoveringTilesDetailsProvider} from './covering_tiles_details_provider';
 
 export class MercatorTransform implements ITransform {
